requirements.txt
numpy==1.21.4
pandas==1.2.4	
Pillow==9.1.0
plotly_express==0.4.1
streamlit==1.1.0
